+++ bin/publish.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env sh
+
+set -e
+
+bin="`dirname "$0"`"
+root="$bin/.."
+
+tag-release "$root"
+create-github-release "$root"
+
+update-github-metadata "$root"